chore: tidy up of some UDAF code
